CRUD de DispositivoIoT listo

+++ TFI_DispositivoIoT_ConfigRed/src/ar/edu/utn/prog2int/Models/ConfiguracionRed.java
@@ -24,7 +24,7 @@ public class ConfiguracionRed extends Base {
         this.dhcpHabilitado = dhcpHabilitado;
     }
 
-    public ConfiguracionRed(int id, boolean eliminado ,String ip, String mascara, String gateway, String dnsPrimario, boolean dhcpHabilitado) {
+    public ConfiguracionRed(Long id, boolean eliminado ,String ip, String mascara, String gateway, String dnsPrimario, boolean dhcpHabilitado) {
         super(id, eliminado);
         this.ip = ip;
         this.mascara = mascara;
